+++ source/dnode/mnode/impl/inc/mndInt.h
@@ -70,7 +70,7 @@ typedef struct {
 
 typedef struct SMnode {
   int32_t           dnodeId;
-  int32_t           clusterId;
+  int64_t           clusterId;
   int8_t            replica;
   int8_t            selfIndex;
   SReplica          replicas[TSDB_MAX_REPLICA];
